Filter constraints to available devices
When a device does not have cam/mic it does not make sense to request
the constraint for it. This also is often harmful as the constraint
cannot be met and the whole gUM call fails. This change checks if the
device has at least one audio or video input and filters the requested
gUM constraints if no matching device is available.