Add support to add/remove tracks on the fly
New kwmks supports add and remove track. This change adds support for it
and uses this behavior for mute/unmute of local media. The new behavior
is enabled by default and will remove and add back tracks for local
media when they are muted/unmuted.

As a side effect this also releases the cam/mic hardware back to the
operating system. Lights go off, bandwidth is saved, world is ours.