Implement stream playback reset after track add
Whenever a track gets added to the stream already playing back, the
the playbackelement needs to be reset to start playing back the new
track instantly. Also when all tracks have been removed from a stream
the stream becomes unusable and needs to be replaced once new tracks are
added again.