Add standby mode on non-mobile
This adds a standby mode which is the default when the browser is not
detected as mobile browser. The standby mode does not enable and show
cam and mic. Instead it just sits there waiting for incoming or outgoing
calls or for the user to switch to any of the other modes.