Add transport send and receive bytes via stats
To get a better understand of the bandwidths and data used by Meet this
change adds stats gathering for all active peer connections (gathered
all 5 seconds). In and outbound bandwidth and overall data usage is
shown as overlay at the bottom left when in a call and controls are
visible.