Generate manifest icons from app icon
For proper support of the target environments, a bunch of bitmaps
representing the application icon need to be generated. This change adds
a Makefile which generates the required bitmaps. The resulting bitmaps
are commited as part of the source tree to avoid pulling in the
dependencies which are required to create them. Whenever the app icon is
changed, the corresponding make targed in src/images has to be called
manually to refresh the bitmaps.