Skip waiting for service worker
The new service worker in c-r-a 2 uses Google's Workbox. By default
service workers do wait until all pages controlled by that worker are
closed. A reload does not close the page. See the Waiting lifecycle of
service workers at
https://developers.google.com/web/fundamentals/primers/service-workers/lifecycle#waiting.
This change makes the service worker always immediately skop waiting.
Ultimatively this should be extended to a feature in kpop which only
skips waiting when the user actually clicked the button or something
like that.