Use local workbox
After the project was updated to c-r-a, Googles workbox is generating the
service worker. C-r-a currently does not offer any way to specify that
workbox should not be loaded from Google's CDN. This change adds a
script i made a while ago which post-processes the generated service
worker to load workbox from the local static files and also ensures that
the workbox files are made available there.