Identify to kwm with ID token
This change removes the direct dependency on kc identities by providing
the ID token to KWM server instead of the specific user id. KWM server
can handle this in the same way as it was assuming the identity is from
a kc backend while allowing other backends at the same time.