Use server side contacts search if more than 100 contacts
Since client side search only works reliably with a small data set, this
change adds server side search. If client or server is automatically
detected by probing if there are more than 100 contacts in the servers
addressbook. If not, all are synced to the client and the behavior is
unchanged. If there are more, none are synced and contacts will be
searched on the server whenever input is entered into the search field.