Add more flexibility to guest mode
This change passes all parameters from the guest settings to the logon
API allowing more future flexibility without further changes. Also in
addition to allowing public rooms, also the token alone can now be used
to try guest logon. It is left to the logon API to decide wether or not
a guest may pass.