Use kwm profile displayName as stream displayName
KWM might be able to provide a displayName for streams together with
incoming and outgoing call events. This change uses that value if a
stream has no displayName from another source already. This is
especially helpful when the client runs as guest without any contacts or
when a peer is no contact.