Always request all scopes, even for guests
To allow normal sign-in to succeed even i guest mode, this change uses
the original OIDC scopes extended by the guest scope, instead of a
reduced set of hard coded scopes. This fixes the problem that signed in
users clicking on guest links did not see their contacts or recents
since they did not get the scopes to access that.