Remove depcreated inactive event in favour of onended
To detect that a screen sharing stream has ended, the inactive event was
used. This event was removed from the spec and thus caused issues in
Firefox which does not trigger it. This change uses the onended event
instead and manually triggers it when screen sharing is stopped by the
user. This fixes a problem when Firefox did not properly end screen
sharing and also takes care of detecting stream stops which are caused
by browser ui elements outside of the app itself.