Decode URL paths for guest access
This change fixes an issue where it was not possible to call into groups
with characters in the name which need encoding in an URL. Since the
group name is not encoded it must be written decoded in the access token
to allow the match check to pass. A call to decodeURI is introduced when
Meet creates the passthru data to be added to the token.