Avoid hydrating device IDs when using Safari
Media device ids are stored presistently on the device so that the
previously configured devices are automatically use when the app
is openend. Since Safari randomizes the device IDs, this approach
does not work. Furthermore Safari also fails to show the permission
prompt when no matching device is found. For the time being this change
does not rehydrate the device settings when Safari is detected.