More ui for syninc user media state with browser
In certain cases the device might need a while to find available media
devices (looking at you Firefox) and feedback is required to indicate to
the user that something is going on. This change adds corresponding
loading animations to the get user media and get display media actions
and applies additional logic to media state handling to fix varios edge
cases when the browsers permission ui kicked in.