Update to kwmjs 1.0.1
It was noticed that peer connections (especially for screen sharing)
sometimes did not connect and the screen sharing area stayed grey
(showing the background).

After investigation i turned out that remote candidates were not applied
to the browser in a certain case. A new kwmjs was released with a fix
for the issue and this change updates it for Meet.