Use correct API prefix default for grapi API
I noticed that the fallback default value for grapi was wrong. Seems a
copy/paste error has found its way to a release. The wrong value is used
whenever a config.json does not include the `apiPrefix` key, leading to
failure for all grapi requests due to wrong URL. This changes the URL to
the correct value, allowing empty configurations or configurations
without `apiPrefix` to work.