Refactor callview actions
To get more control of the apps behavior, independent from the currently
showing view, this change refactors a lot of actions out of the
component. The initial reason for this change was to get reliable
behavior when auto join is enabled on application start (which did or
did not work depending on how fast kwm was connected). With this change,
the auto join logic waits until it actually can join successfully.