Improve audioSink apply logic
I noticed that the audio sink setting logged AbortError quite often.
After investigation, this turned out to happen when the audio sink is
set multiple times. This change memoizes setting the audio sink so it is
only set once to the same element. This still has the potential for
aborts when the audio sink is changed very quickly but i guess that is
ok since the app does not really offer a way to do that.