Disable session monitor for guest users
Guest sessions are bound to a request token created by kwm server. This
means the guest session is always active when this token is present. As
a result the session monitor cannot detect session changes since it
always checks the guest session provided via token. I noticed that this
can trigger a session monitor storm if multiple guest sessions are
active in the same context. Since the session monitor is useless for
guests in the first place, this change stops the session monitor
whenever a guest user is detected.