Offer audio unmute when auto play policy prevents play
Modern browsers have auto play policies. With this change, all incoming
streams are muted by default. This means they always auto play. In
addition the app now checks wether auto play is possible and then
enabled streams. If not possible, a snack is shown which notifies the
user and offers the option to enable audio. Since that is a user
interaction it also automatically allows audio to play afterwards.