Improve default guest behavior
When guest mode is enabled by default, it was required to provide the
guest parameter in the hash. With this change public channels can be
joined without this parameter. Instead if the parameter is not present
the default will be used and the current channel path from the URL is
added, using both to attempt guest sign in. If successfull the
designated channel is joined as guest using the default mode.

Fixes: KWM-87