Hide peers without video in call grid for groups
In groups, this change displays only remote peers which are actually
sending a video stream. Audio only peers will not occupy space and are
only visible in the participant list in the left pane.

Currently, this means that the indication if the remote peer is sending
audio or not is lost and will be brought back in a later commit.