Add indicator for audio and video status to participants
Streams are classified wether or not they contain audio and/or video.
This classification is now made available in the store and thus becomes
accessible for all components. It is added to the participants list so
it becomes clear which participants send a given stream even when no
video of such a participant is shown. For the local user, the
classification is taken from the media mute status.