Resolve a flex issue for older browsers
Older browsers show flex issues which could lead to video in mobile view
getting rendered full size instead of giving space to other UI elements.
A workaround is implemented in this change which resolves the issue for
Samsung Internet Browser (current version at time of writing is based on
Chromium 83).